Fix crash on slow network connections
The crash occurred after entering the card number if the configuration
had not finished loading. Drop-in tried to auto-advance to the next
field, but if the configuration hadn't finished loading, we didn't
have the list of required fields yet, so there was nothing to advance to.
The fix for this is to check if the index of the current field in the
list of required fields is NSNotFound, and if so, skip the attempt to
advance to the next field.

Additionally, we were copying BTAPIClient for each view controller that
subclasses BTDropInBaseViewController. This meant that each view
controller had to fetch the configuration, instead of using the cached
configuration previously fetched. If we use the same instance of
BTAPIClient across view controllers, we'll reduce latency by using the
cached configuration and avoid any race conditions (like the one above).